chore: renaming module elasticsearch -> backend
