+++ src/rubrix/server/backend/mappings/text2text.py
@@ -1,4 +1,4 @@
-from rubrix.server.elasticseach.mappings.helpers import mappings
+from rubrix.server.backend.mappings.helpers import mappings
 
 
 def text2text_mappings():
