+++ src/rubrix/server/backend/mappings/text_classification.py
@@ -1,4 +1,4 @@
-from rubrix.server.elasticseach.mappings.helpers import mappings
+from rubrix.server.backend.mappings.helpers import mappings
 
 
 def text_classification_mappings():
