+++ src/rubrix/server/backend/mappings/token_classification.py
@@ -2,8 +2,8 @@ from typing import Any, Dict, Optional
 
 from pydantic import BaseModel, Field
 
-from rubrix.server.elasticseach.mappings.helpers import mappings
-from rubrix.server.elasticseach.query_helpers import nested_mappings_from_base_model
+from rubrix.server.backend.mappings.helpers import mappings
+from rubrix.server.backend.query_helpers import nested_mappings_from_base_model
 
 
 class MentionMetrics(BaseModel):
