+++ src/rubrix/server/backend/metrics/text_classification.py
@@ -2,11 +2,8 @@ import dataclasses
 from typing import Any, Dict, List, Optional
 
 from rubrix.server._helpers import unflatten_dict
-from rubrix.server.elasticseach.metrics.base import (
-    ElasticsearchMetric,
-    TermsAggregation,
-)
-from rubrix.server.elasticseach.query_helpers import filters
+from rubrix.server.backend.metrics.base import ElasticsearchMetric, TermsAggregation
+from rubrix.server.backend.query_helpers import filters
 
 
 @dataclasses.dataclass
