+++ src/rubrix/server/backend/search/model.py
