+++ tests/functional_tests/search/test_search_service.py
@@ -9,8 +9,8 @@ from rubrix.server.apis.v0.models.text_classification import (
     TextClassificationRecord,
 )
 from rubrix.server.apis.v0.models.token_classification import TokenClassificationQuery
+from rubrix.server.backend.elasticsearch import ElasticsearchBackend
 from rubrix.server.daos.records import DatasetRecordsDAO
-from rubrix.server.elasticseach.backend import ElasticsearchBackend
 from rubrix.server.services.metrics import MetricsService
 from rubrix.server.services.search.model import SortConfig
 from rubrix.server.services.search.service import SearchRecordsService
